fix: set type dir
