+++ src/@types/data.type.d.ts
@@ -1,0 +1,14 @@
+declare module 'data-type' {
+  export type InfoFormData = {
+    name: string;
+    id: string;
+    phone: string;
+    email: string;
+  };
+
+  export type ProfileData = {
+    name: string;
+    src: string;
+    user: string;
+  };
+}
