--- src/components/_common/Btn.d.ts
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export type BtnProps = {
-  label: string;
-  styleClass: string;
-  onClick: (e: React.MouseEvent<HTMLElement, MouseEvent>) => void;
-};
