--- src/components/_common/Title.d.ts
@@ -1,3 +1,0 @@
-export type TitleProps = {
-  label: string;
-};
