--- src/components/mainpage/Content.d.ts
@@ -1,7 +1,0 @@
-export type ContentProps = {
-  svg: string;
-  title: string;
-  content: string;
-  styleClass: string;
-  subtitle?: string;
-};
